Initial implementation
I wrote this as a single-file script in a different repo but then
decided to break it out into a proper Clojure project in its own repo.

This is the same code that was in the single-file script, more or less,
but broken into multiple namespaces. There’s still some work to do
there; `core` is still too big.

Refs FundingCircle/docs#237