Replace Python API client with Clojure code
This Python library was great for bootstrapping the tool, but the GIL is
making the publishing process way slow.

Honestly, now that I think about it, I don’t think it’s *really* the
GIL. Because we could have used threading in Python to achieve
concurrency without parallelism, and that would have been sufficient,
since our publishing process spends most of its time waiting for
responses from Confluence.

However, while that would have been *technically* feasible, it would
have been too much work and it would have made the code hard to
maintain. That’s because we really need the concurrency to happen at a
higher level than the HTTP requests. We need a way for the core ns to
submit a ton of upserts simultaneously and block while waiting for the
results. And upserts are complex workflows involving multiple calls to
Confluence.

Now, yes, I could have added a Python ThreadPoolExecutor to the upsert
function, and have it return Python futures or something like that, I
guess. Or maybe Python futures wrapped in Clojure futures. But it would
have been awkward to implement, difficult to reason about, and I had
other reasons to remove the usage of Python via libpython-clj from the
project. Namely, I’m hoping to distribute this tool by compiling it to a
native binary using GraalVM. That’s a tricky enough prospect at this
juncture; it’d be effectively impossible (for me) with Python — and a
Python library (the Atlassian library) being required for the project.

Come to think of it, removing the dependency on Python and the Python
library will also make development and testing simpler and easier, as
there will be less requirements and steps involved in setting up the
environment.

Resolves #2