Configure clj-kondo (for linting Clojure code)
Until now the default configuration has worked fine, but in the prior
commit we started using the library Claypoole, and clj-kondo doesn’t
understand how to lint one of its macros. This config teaches it how.

And as long as I was at it, I enabled a few more linters.

https://github.com/borkdude/clj-kondo/