TE efficiency tweak
I like the efficiency improvement here (from checking whether the
existing page with the same title in the space has a different parent,
which is an error condition (but also an edge case)) but it does shift
`upsert` from a little too long to way too long, especially for an edge case. It’ll need a little more refactoring.