Fix a bug in link resolution caused by README integration
This walks back most of the changes made in c69a0aeb0d which turned
out to cause problems for link resolution. This is explained in the
diff of this commit in the expanded docstring of `integrate-readme`.

Now, instead of removing READMEs from the page-tree — which
impedes link resolution — we’re retaining them, and filtering them
out when publishing.