Streamline our GHA workflow “Test”
GHA = GitHub Actions

In #26, wherein I first added a GHA workflow, @robertstettner commented:

> Looking at the workflow, it is clear to me that GitHub Actions for
> CI/CD is not easy, and looks complex.

My impression, based on his comments in #26 and #27, is that he feels
that Drone pipelines are better somehow.

I might be mistaken, but my impression is that this feeling is based on
the pipeline/workflow definition files. And I must admit, the definition
file that I introduced in #26 is overly long and hard to read.

So that we can more accurately compare and contrast GHA and Drone, I
thought I’d see whether I could make the GHA workflow clearer and more
concise. I think I succeeded.

(I had to use a pre-release version of the action `actions/cache`, but
it seems to work well. The design of GHA incentivizes the maintainers of
actions to only increment the major version number when they have either
a breaking change or some massive new featureset that they want to
encourage people to switch to. The team that maintains `actions/cache`
has a few more features they want to ship before they release 2.0. 🤷‍♂️)