Don’t run empty test suites in GHA
GHA = GitHub Actions

I included these empty test suites in 22cb2b18 as an experiment, or
maybe a proof-of-concept… I wanted to try to really see the power of
matrix jobs. Now that I’ve done that, it doesn’t really make sense to
keep running empty test suites. So this removes them from the matrix,
for now.