Fix the bug in get-page-space-key
As described in the prior commit, which added a failing test for this
bug.

It just occurred to me that rather than fixing this bug, we probably
should have simplified the code. See, the only reason we need a cache is
because we have to keep looking up the space keys for parent pages —
right now, every single time `upsert` is called. This is, I guess,
because I modeled the API for `upsert` from a generic Confluence API
wrapper library. In our case, however, we can declare that all calls to
`upsert` can and should use the same space key. This will be simpler,
faster, more efficient, more resilient, and less likely to cause bugs.
So we should do that. Let’s instead require that the space key be passed
to `upsert` and let the caller worry about getting it and storing it.
Maybe in the calling context we’ll shove it into the `client` so we
don’t have yet another thing to keep passing around everywhere.