Speed up publishing by adding stupid concurrency
For way too many days I’ve been working on making publishing concurrent
the “right” way with ExecutorServices, thread pools, futures, etc. And
I keep running into situations that seem to cause deadlocks. I don’t
seem to be smart enough to reason about concurrency and blocking when
there’s recursion involved.

Just now I somehow had the epiphany that a stupid approach to
concurrency would work for my team, for right now, because I could apply
the concurrency to the direct children of the root only, and we
currently have 13 or 14 of those. That means that even though `pmap`
uses an unbounded thread pool, which is why I’d been avoiding it for
calling Confluence, in our case this should result in no more than 13–14
concurrent requests to Confluence. That’s higher than I’d like — I was
going to use 10 as my initial hard-coded value — but our Confluence
server seems to be able to handle it OK. So what the hell. Let’s do the
stupidest, simplest thing possible, for as long as it’s viable.

Closes #8 I guess.