Add exception handling to more calls to Confluence
I was testing this locally and an exception bubbled out of this
namespace, aborting my entire workflow. That helped me realize that I’d
missed a few calls when I added exception handling in the prior commit.

I’ll test this in a few minutes. It’ll probably work fine, but the
end-result of the two commits combined has me rethinking the approach.
It’s a lot of code, and it harms readability and thus maintainability.

And I think I see a potential simpler approach: to change the hato
config to throw exceptions, and then we can handle *all* failures the
same way: by converting the exception to a fault. That’d be simpler than
the current approach, wherein many of the operations have two different
error-handling pathways: one for a request that wasn’t successful but no
exception was thrown, and one for exceptions (e.g. timeouts, DNS
failures, etc.)