Minor code improvement from code review
While reviewing #32, @MattSouthrden and I realized that the prior
commit, while reasonable, had created a problem: it was now possible to
call `make-client` and get back a client that is not usable for certain
operations (`create-page` and `upsert`) because that client wouldn’t
have the ::space-key kvp. We didn’t like this, it seemed unintuitive and
problematic for testing.

So this changes the API such that a caller must now pass the space-key
directly to `create-page` and `upsert` as args, and it’s no longer added
to the client map at all. This is slightly more verbose but it’s clearer
and simpler and more intuitive (in the sense of the Principal of Least
Surprise).

Co-authored-by: Matt Southerden <matt.southerden@fundingcircle.com>